% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_functions.R
\name{MPT.Corr}
\alias{MPT.Corr}
\title{MULTIVARIATE PERMUTATION TESTING OF CORRELATIONS}
\usage{
MPT.Corr(observed.data, criterion, predictor.start, predictor.end,
  output.file = "Results.txt", test = 2, tail = 2, alpha = 0.05,
  output = 0, permutations = 10000)
}
\description{
"This R program performs multivariate permutation tests of the correlation between a single criterion variable and multiple predictor variables. Data must be imported by user with subjects/cases as rows and variables as columns.  Missing data must be specified."
This script was Written by Jennifer Urbano Blackford, Ph.D. of The Kennedy Center, Vanderbilt University, Nashville, TN.
The original can be found at http://users.cla.umn.edu/~nwaller/downloads/mpt/mptcorr.r
}
