% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atf_functions.R
\name{read.atf}
\alias{read.atf}
\title{Read ATF}
\usage{
read.atf(input.file.name = "short_sine.atf",
  input.dir.path = "S:/Data_Daniel/test_atf_write/")
}
\description{
This function takes an ATF, strips the header and returns a dataframe for easy visualization or manipulaiton.
It's simply a wrapper for read.table. Ensure that the path ends with a slash.
}
