% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Theme Publication}
\usage{
theme_Publication(base_size = 14, base_family = "Arial")
}
\description{
This theme to clean up ggplots for presentations.
}
