% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atf_functions.R
\name{write.to.atf}
\alias{write.to.atf}
\title{Write ATF}
\usage{
write.to.atf(input.data.frame = as.data.frame(cbind(1:100, sin(1:100))),
  output.dir.path = "S:/Data_Daniel/test_atf_write/",
  output.file.name = "short_sine.atf")
}
\description{
This function takes a two column dataframe and saves it as an Axon Text File.
Ensure that the path ends with a slash.
}
